渠道活动（fix）：活动上下架状态修改异常

+++ src/main/java/com/xiliulou/electricity/service/impl/ChannelActivityServiceImpl.java
@@ -134,8 +134,8 @@ public class ChannelActivityServiceImpl implements ChannelActivityService {
         }
         
         Integer tenantId = TenantContextHolder.getTenantId();
-        
-        if (!Objects.equals(tenantId, channelActivity.getTenantId())) {
+    
+        if (!Objects.equals(tenantId, channelActivity.getTenantId().intValue())) {
             return Triple.of(true, "", "");
         }
